# workbenchtimetrackerapp.client/src/app/api/model/personDTO.ts
/**
 * WorkbenchTimeTrackerApp.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkTaskDTO } from './workTaskDTO';


export interface PersonDTO { 
    id?: number;
    fullName: string;
    workTasks?: Array<WorkTaskDTO>;
}