# workbenchtimetrackerapp.client/src/app/api/model/workTaskDTO.ts
/**
 * WorkbenchTimeTrackerApp.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimeEntryDTO } from './timeEntryDTO';


export interface WorkTaskDTO { 
    id?: number;
    name: string;
    description?: string;
    timeEntries?: Array<TimeEntryDTO>;
}